Codes/experiments/rbs/experiment_rbs_diffusion_ga_variant3.py
# -*- coding: utf-8 -*-
# 文件名: main_model3.py
import sys
import os
import random
import torch
import torch.nn as nn
import torch.nn.functional as F
import torch.optim as optim
from torch.utils.data import TensorDataset, DataLoader
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from scipy.spatial.distance import jensenshannon
from scipy.linalg import sqrtm
from collections import Counter
import logomaker
from tqdm import tqdm
import math
import Levenshtein
import gc


# ==============================================================================
# 第0部分: 环境设置
# ==============================================================================
def set_seed(seed=42):
    random.seed(seed)
    np.random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)
    torch.backends.cudnn.deterministic = True
    torch.backends.cudnn.benchmark = False


set_seed(42)
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
output_dir = "output_plots_rigorous_v3"
os.makedirs(output_dir, exist_ok=True)

# ==============================================================================
# 第1部分: 模型架构 (UNet & Oracle)
# ==============================================================================
SEQUENCE_LENGTH = 17
NUCLEOTIDES = 4
CSV_FILE = "rbs_data.csv"


class TimeEmbedding(nn.Module):
    def __init__(self, dim):
        super().__init__()
        self.dim = dim

    def forward(self, t):
        half_dim = self.dim // 2
        embed = np.log(10000) / (half_dim - 1)
        embed = torch.exp(torch.arange(half_dim, device=t.device) * -embed)
        embed = t[:, None] * embed[None, :]
        embed = torch.cat((embed.sin(), embed.cos()), dim=1)
        return embed


class AttentionBlock(nn.Module):
    def __init__(self, n_channels):
        super().__init__()
        self.norm = nn.GroupNorm(8, n_channels)
        self.qkv = nn.Conv1d(n_channels, n_channels * 3, 1)
        self.out = nn.Conv1d(n_channels, n_channels, 1)

    def forward(self, x):
        batch, channel, length = x.shape
        x_norm = self.norm(x)
        qkv = self.qkv(x_norm).view(batch, 3, channel, length)
        q, k, v = qkv.unbind(1)
        attn = torch.einsum('bcl,bck->blk', q, k) * (channel ** -0.5)
        attn = attn.softmax(dim=-1)
        out = torch.einsum('blk,bck->bcl', attn, v)
        return x + self.out(out)


class ResidualBlock(nn.Module):
    def __init__(self, in_channels, out_channels, time_emb_dim):
        super().__init__()
        self.mlp_time = nn.Sequential(nn.SiLU(), nn.Linear(time_emb_dim, out_channels))
        self.conv1 = nn.Conv1d(in_channels, out_channels, kernel_size=3, padding=1)
        self.norm1 = nn.GroupNorm(8, out_channels)
        self.act1 = nn.SiLU()
        self.conv2 = nn.Conv1d(out_channels, out_channels, kernel_size=3, padding=1)
        self.norm2 = nn.GroupNorm(8, out_channels)
        self.act2 = nn.SiLU()
        self.shortcut = nn.Conv1d(in_channels, out_channels, 1) if in_channels != out_channels else nn.Identity()

    def forward(self, x, t):
        h = self.act1(self.norm1(self.conv1(x)))
        h += self.mlp_time(t)[:, :, None]
        h = self.act2(self.norm2(self.conv2(h)))
        return h + self.shortcut(x)


class UpDownBlock(nn.Module):
    def __init__(self, in_channels, out_channels, time_emb_dim, down=True):
        super().__init__()
        self.res = ResidualBlock(in_channels, out_channels, time_emb_dim)
        self.attn = AttentionBlock(out_channels)
        self.sample = nn.Conv1d(out_channels, out_channels, 3, stride=2, padding=1) if down else \
            nn.ConvTranspose1d(out_channels, out_channels, 4, stride=2, padding=1)

    def forward(self, x, t):
        return self.sample(self.attn(self.res(x, t)))


class UNet(nn.Module):
    def __init__(self, n_channels=64):
        super().__init__()
        time_emb_dim = n_channels * 4
        self.time_embedding = nn.Sequential(TimeEmbedding(n_channels), nn.Linear(n_channels, time_emb_dim), nn.SiLU(),
                                            nn.Linear(time_emb_dim, time_emb_dim))
        self.in_conv = nn.Conv1d(NUCLEOTIDES, n_channels, 3, padding=1)
        self.down1 = UpDownBlock(n_channels, n_channels * 2, time_emb_dim, True)
        self.down2 = UpDownBlock(n_channels * 2, n_channels * 4, time_emb_dim, True)
        self.mid = ResidualBlock(n_channels * 4, n_channels * 4, time_emb_dim)
        self.up1 = UpDownBlock(n_channels * 8, n_channels * 2, time_emb_dim, False)
        self.up2 = UpDownBlock(n_channels * 4, n_channels, time_emb_dim, False)
        self.out = nn.Conv1d(n_channels * 2, NUCLEOTIDES, 1)

    def forward(self, x, t):
        t_emb = self.time_embedding(t)
        x_in = self.in_conv(x)
        x1 = self.down1(x_in, t_emb)
        x2 = self.down2(x1, t_emb)
        xm = self.mid(x2, t_emb)
        u1 = self.up1(torch.cat([xm, x2], 1), t_emb)
        u2 = self.up2(torch.cat([F.interpolate(u1, x1.shape[2]), x1], 1), t_emb)
        return self.out(torch.cat([F.interpolate(u2, x_in.shape[2]), x_in], 1))


class OracleCNN(nn.Module):
    def __init__(self):
        super().__init__()
        self.conv = nn.Sequential(nn.Conv1d(4, 64, 7, padding=3), nn.ReLU(), nn.MaxPool1d(2),
                                  nn.Conv1d(64, 128, 5, padding=2), nn.ReLU(), nn.MaxPool1d(2), nn.Flatten())
        self.fc = nn.Sequential(nn.Linear(128 * (SEQUENCE_LENGTH // 4), 64), nn.ReLU(), nn.Linear(64, 1))

    def forward(self, x): return self.fc(self.conv(x))

    def get_embedding(self, x): return self.fc[0](self.conv(x))


# ==============================================================================
# 第2部分: 数据严谨划分与评估指标
# ==============================================================================
def calculate_novelty(gen_strs, train_strs, n_check=500):
    """计算生成序列与训练集的最短编辑距离 (衡量新颖性)"""
    sample_train = random.sample(train_strs, min(len(train_strs), 2000))
    novs = [min([Levenshtein.distance(g, t) for t in sample_train]) for g in gen_strs[:n_check]]
    return np.mean(novs)


print("--- [main_model3.py] 数据预处理与严谨划分 ---")
char_map = {'A': 0, 'C': 1, 'G': 2, 'T': 3}
idx_to_char = {v: k for k, v in char_map.items()}

if not os.path.exists(CSV_FILE):
    df = pd.DataFrame({'序列': ["".join(np.random.choice(list(char_map.keys()), 17)) for _ in range(10000)],
                       'rl': np.random.rand(10000)})
else:
    df = pd.read_csv(CSV_FILE).head(260000)


def encode(s):
    res = np.zeros((4, 17), dtype=np.float32)
    for i, c in enumerate(s.upper()): res[char_map[c], i] = 1.0
    return res


one_hot_all = np.array([encode(s) for s in df['序列']])
scores_all = df['rl'].values.astype(np.float32)
all_strs = df['序列'].tolist()

# 严谨划分: Train (80%), Val (10%), Test (10%)
X_train, X_temp, y_train, y_temp, str_train, str_temp = train_test_split(one_hot_all, scores_all, all_strs,
                                                                         test_size=0.2, random_state=42)
X_val, X_test, y_val, y_test, str_val, str_test = train_test_split(X_temp, y_temp, str_temp, test_size=0.5,
                                                                   random_state=42)

# 训练 Oracle (带早停)
real_oracle = OracleCNN().to(device)
opt_o = optim.Adam(real_oracle.parameters(), lr=1e-3)
best_v = float('inf')
print("训练 Oracle 基准 (含验证集监控)...")
for e in range(30):
    real_oracle.train()
    for b in range(0, len(X_train), 256):
        opt_o.zero_grad()
        loss = F.mse_loss(real_oracle(torch.from_numpy(X_train[b:b + 256]).to(device)).squeeze(),
                          torch.from_numpy(y_train[b:b + 256]).to(device))
        loss.backward();
        opt_o.step()
    real_oracle.eval()
    with torch.no_grad():
        v_l = F.mse_loss(real_oracle(torch.from_numpy(X_val).to(device)).squeeze(),
                         torch.from_numpy(y_val).to(device)).item()
        if v_l < best_v: best_v = v_l; torch.save(real_oracle.state_dict(), "best_oracle.pth")
real_oracle.load_state_dict(torch.load("best_oracle.pth"))


# ==============================================================================
# 第3部分: 采样与辅助函数
# ==============================================================================
def one_hot_to_strings(oh):
    idx = torch.argmax(oh, dim=1).cpu().numpy()
    return ["".join([idx_to_char[i] for i in s]) for s in idx]


def generate_sequences_in_batches(model, oracle, n, bs, guidance=0.5, T=1000):
    model.eval();
    res = []
    betas = torch.linspace(1e-4, 0.02, T, device=device)
    alphas = 1. - betas
    alphas_cum = torch.cumprod(alphas, 0)
    for _ in range(math.ceil(n / bs)):
        curr_bs = min(bs, n - len(res) * bs)
        x = torch.randn((curr_bs, 4, 17), device=device)
        for t in reversed(range(T)):
            t_batch = torch.full((curr_bs,), t, device=device)
            if guidance > 0:
                with torch.enable_grad():
                    x.requires_grad_(True)
                    g = torch.autograd.grad(oracle(x).sum(), x)[0].clamp(-1, 1)
            else:
                g = 0
            with torch.no_grad():
                eps = model(x, t_batch.float()) - torch.sqrt(1 - alphas_cum[t]) * g * guidance
                x = (x - (1 - alphas[t]) / torch.sqrt(1 - alphas_cum[t]) * eps) / torch.sqrt(alphas[t])
                if t > 0: x += torch.sqrt(betas[t]) * torch.randn_like(x)
        res.append(x.cpu())
    return torch.cat(res, 0).to(device)


# ==============================================================================
# 第4部分: 主评估流程 (Nature 风格对标)
# ==============================================================================
if __name__ == "__main__":
    # 此处假设已经加载了训练好的最优 UNet (T=1000, CH=128)
    unet = UNet(n_channels=128).to(device)
    # 模拟评估 (实际应加载 .pth)

    print("\n--- [Part 5] 严谨性基准测试 ---")
    N_EVAL = 500

    # 采样
    gen_oh = generate_sequences_in_batches(unet, real_oracle, N_EVAL, 64, guidance=1.0)
    gen_str = one_hot_to_strings(gen_oh)

    # 指标计算
    nov_score = calculate_novelty(gen_str, str_train)
    with torch.no_grad():
        test_scores = real_oracle(torch.from_numpy(X_test[:N_EVAL]).to(device)).cpu().numpy().flatten()
        gen_scores = real_oracle(gen_oh).cpu().numpy().flatten()

    print(f"统计指标 | Test Set Avg: {np.mean(test_scores):.4f} | Generated Avg: {np.mean(gen_scores):.4f}")
    print(f"生成新颖性 | 平均编辑距离 (vs Train): {nov_score:.2f} (证明非背诵)")

    # 绘图: Violin Plot
    plt.figure(figsize=(8, 6))
    sns.violinplot(data=[test_scores, gen_scores], palette="pastel")
    plt.xticks([0, 1], ["Experimental (Test)", "Diffusion (Generated)"])
    plt.title("Distribution of Functionality Scores")
    plt.savefig(os.path.join(output_dir, "rigorous_comparison.png"), dpi=300)
    plt.show()