requirements.txt
# Human 5' UTR modeling - dependencies
# Python 3.8+ recommended

# Deep learning
torch>=1.13.0

# Numerical & data
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
logomaker>=0.8

# Utilities
tqdm>=4.60.0
python-Levenshtein>=0.12.0
